Makefile: Add non-Kbuild portion, using KDIR to specify kernel location
Extend the Makefile with a non-Kbuild portion to simplify things for the
user.

To build for the current (running) kernel, the command is simply:

    make

or:

    make modules

Then to install for the current kernel:

    sudo make install

or:

    sudo make modules_install

To build and install for a specific kernel, the `KDIR` variable can be
set to specify the path to the kernel header directory, which is the same
directory as used with the `-C` option previously.

For example, to build for a specific kernel:

    make KDIR=/usr/src/linux-headers-5.15.84-v8+

or:

    make KDIR=/usr/src/linux-headers-5.15.84-v8+ modules

Then to install for the specific kernel:

    sudo make KDIR=/usr/src/linux-headers-5.15.84-v8+ install

or:

    sudo make KDIR=/usr/src/linux-headers-5.15.84-v8+ modules_install

Note: The `KDIR` value should be the same for building and installing!